# openapi/os-api.yaml
openapi: 3.1.0
info:
  title: Shennong OS API
  version: 1.0.0
  description: |
    V1 control-plane API for Shennong OS. Unscoped Resource catalog endpoints
    expose only Resources whose permissions.visibility is exactly public.
    Private Resources are available only through an authenticated Project scope
    with an explicit binding and Project role.

    Authenticated mutations require an allowed `Origin`. Cookie-authenticated
    mutations additionally require `X-CSRF-Token` matching the `shennong_os_csrf`
    cookie. Bearer tokens are opaque Shennong OS session tokens, not JWTs.

    Shennong OS is the sole Project and RBAC authority. Project-scoped data-plane
    calls first reconcile an idempotent, non-authoritative Research Project
    shadow in headless ShennongDB; DB never grants access from that shadow.
servers:
  - url: /
...
security:
  - sessionCookie: []
  - bearerSession: []
paths:
...
  /api/v1/setup/status:
    get:
      tags: [Setup]
      operationId: getSetupStatus
      security: []
      responses:
        '200':
          description: Whether the first administrator must be created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetupStatusEnvelope'
        default: { $ref: '#/components/responses/Error' }
  /api/v1/setup/admin:
    post:
      tags: [Setup]
      operationId: bootstrapAdministrator
      description: |
        Creates the first and only bootstrap administrator. The server
        serializes competing requests with a PostgreSQL advisory transaction
        lock and rechecks that no user exists before its insert.
      security: []
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/BootstrapToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountRequest' }
      responses:
        '201':
          description: First administrator created and session cookies issued.
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionEnvelope' }
        '409': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/registration-policy:
    get:
      tags: [Authentication]
      operationId: getRegistrationPolicy
      security: []
      responses:
        '200':
          description: Current registration mode.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistrationPolicyEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      operationId: registerUser
      description: Creates an ordinary user. An invitation is required in invite-only mode.
      security: []
      parameters:
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountRequest' }
      responses:
        '201':
          description: User created and session cookies issued.
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/sign-in:
    post:
      tags: [Authentication]
      operationId: signIn
      security: []
      parameters:
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignInRequest' }
      responses:
        '200':
          description: Session created.
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/sign-out:
    post:
      tags: [Authentication]
      operationId: signOut
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204': { description: Session revoked and cookies expired. }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/session:
    get:
      tags: [Authentication]
      operationId: getSession
      responses:
        '200':
          description: Current user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentSessionEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/sessions:
    get:
      tags: [Authentication]
      operationId: listOwnSessions
      responses:
        '200':
          description: Active opaque sessions for the current user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/auth/sessions/{id}:
    delete:
      tags: [Authentication]
      operationId: revokeOwnSession
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204': { description: Session revoked. }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/admin/registration-policy:
    patch:
      tags: [Administration]
      operationId: updateRegistrationPolicy
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegistrationPolicyWrite' }
      responses:
        '200':
          description: Policy updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RegistrationPolicyEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/admin/invites:
    get:
      tags: [Administration]
      operationId: listInvitations
      responses:
        '200':
          description: Invitation metadata. Raw codes are never returned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListEnvelope' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Administration]
      operationId: createInvitation
      description: The raw invitation code is returned only by this response.
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InviteCreate' }
      responses:
        '201':
          description: Invitation created; store its one-time code now.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObjectEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/admin/invites/{id}:
    delete:
      tags: [Administration]
      operationId: revokeInvitation
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204': { description: Invitation revoked. }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/projects:
    get:
      tags: [Projects]
      operationId: listProjects
      responses:
        '200':
          description: Projects visible to the actor.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListEnvelope' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Projects]
      operationId: createProject
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreate' }
      responses:
        '201':
          description: Private project created with the actor as owner.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObjectEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/projects/{id}:
    get:
      tags: [Projects]
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Project details.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObjectEnvelope' }
        default: { $ref: '#/components/responses/Error' }
    patch:
      tags: [Projects]
      operationId: updateProject
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectUpdate' }
      responses:
        '200':
          description: Project updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObjectEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/projects/{id}/members:
    get:
      tags: [Projects]
      operationId: listProjectMembers
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Project members.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/projects/{id}/members/{user_id}:
    put:
      tags: [Projects]
      operationId: putProjectMember
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MemberWrite' }
      responses:
        '200':
          description: Membership created or updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObjectEnvelope' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      tags: [Projects]
      operationId: deleteProjectMember
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204': { description: Membership deleted. }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/projects/{id}/graph/subgraph:
    get:
      tags: [Projects]
      operationId: getProjectGraphSubgraph
      description: Project-RBAC graph traversal. The OS reconciles its authoritative project shadow and requires Shennong DB to enforce the same project boundary.
      parameters:
        - $ref: '#/components/parameters/Id'
        - name: root
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 180
            pattern: '^[A-Za-z0-9._:-]+$'
        - name: depth
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 3, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 80 }
      responses:
        '200':
          description: Project-scoped graph subgraph.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObjectEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/projects/{id}/uploads:
    get:
      tags: [Projects]
      operationId: listProjectUploads
      description: Lists upload records inside the authenticated Project read boundary.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Project-scoped upload records.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListEnvelope' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Projects]
      operationId: uploadProjectFile
      description: Streams one non-empty file to Shennong DB. OS derives the actor and Project UUID from the authenticated session and never accepts either identity from browser input.
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CsrfToken'
        - name: X-Filename
          in: header
          required: true
          schema: { type: string, minLength: 1, maxLength: 255 }
        - name: Content-Length
          in: header
          required: true
          schema: { type: integer, minimum: 1, maximum: 53687091200 }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        '200':
          description: Persisted upload record with server-computed SHA-256.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObjectEnvelope' }
        '411': { description: Content-Length is missing or invalid. }
        '413': { description: Upload exceeds the configured bounded stream size. }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/projects/{id}/uploads/register:
    post:
      tags: [Projects]
      operationId: registerProjectUploads
      description: Atomically creates a private Resource and immutable Artifacts from uploads owned by the same actor and Project, then binds the Resource to that Project.
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [upload_ids, resource_id, name]
              properties:
                upload_ids:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items: { type: string, format: uuid }
                resource_id:
                  type: string
                  minLength: 1
                  maxLength: 160
                  pattern: '^[A-Za-z0-9._-]+$'
                name: { type: string, minLength: 1, maxLength: 512 }
                description: { type: string, maxLength: 4096 }
                organism: { type: string, maxLength: 256 }
                modality: { type: string, maxLength: 256 }
                assay: { type: string, maxLength: 256 }
                reference: { type: string, maxLength: 256 }
                annotation: { type: string, maxLength: 256 }
                format: { type: string, minLength: 1, maxLength: 64, default: binary }
                data_class: { type: string, enum: [raw, canonical, derived], default: raw }
      responses:
        '200':
          description: Private Resource registered and Project-bound.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObjectEnvelope' }
        default: { $ref: '#/components/responses/Error' }
...
  /api/v1/threads/{id}/runs:
    post:
      tags: [Runs]
      operationId: createRun
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RunCreate' }
      responses:
        '201':
          description: Run queued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObjectEnvelope' }
        default: { $ref: '#/components/responses/Error' }
...
  /api/v1/agent/runs/{id}/metadata:
    post:
      tags: [Runs]
      operationId: recordInternalRunMetadata
      security: [{ serviceToken: [] }]
      parameters:
        - $ref: '#/components/parameters/Id'
        - $ref: '#/components/parameters/AgentRuntimeService'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { description: Metadata stored. }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: shennong_os_session
      description: Random opaque session token. Stored server-side only as a SHA-256 hash.
    bearerSession:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-service-token
      description: Agent Runtime callback token loaded by both services from secret files.
    ideSessionCookie:
      type: apiKey
      in: cookie
      name: shennong_ide_session
      description: |
        Opaque, server-hashed access token. It is host-only, HttpOnly,
        SameSite=Strict and path-scoped to exactly one Runtime Session proxy.
...
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...